@@ -23,18 +23,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.2.2",
-    "@angular/cdk": "^13.2.2",
-    "@angular/common": "^13.2.2",
-    "@angular/compiler": "^13.2.2",
-    "@angular/core": "^13.2.2",
-    "@angular/forms": "^13.2.2",
-    "@angular/localize": "^13.2.2",
-    "@angular/material": "^13.2.2",
-    "@angular/platform-browser": "^13.2.2",
-    "@angular/platform-browser-dynamic": "^13.2.2",
-    "@angular/platform-server": "^13.2.2",
-    "@angular/router": "^13.2.2",
+    "@angular/animations": "^14.2.12",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/forms": "^14.2.12",
+    "@angular/localize": "^14.2.12",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/platform-server": "^14.2.12",
+    "@angular/router": "^14.2.12",
     "@elastic/apm-rum-angular": "^2.1.4",
     "@fortawesome/angular-fontawesome": "^0.10.1",
     "@fortawesome/fontawesome-svg-core": "^1.3.0",
@@ -84,10 +84,10 @@
     "zone.js": "~0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.2.3",
-    "@angular/cli": "^13.2.6",
-    "@angular/compiler-cli": "^13.2.2",
-    "@angular/language-service": "^13.2.6",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.12",
+    "@angular/language-service": "^14.2.12",
     "@babel/cli": "^7.18.10",
     "@babel/core": "^7.17.4",
     "@babel/preset-env": "^7.16.11",
